PulseAlgorithm.jl

PulseAlgorithm.jl

Introduction

Installation

Graph

Data Loader

Util

PA-S-$\alpha$RP

PA-SD_RSPP

Introduction

PulseAlgorithm.jl provides an implementation for the Pulse Algorithm for Reliable Shortest Path Problems (PA-RSPPs) designed to tackle challenging stochastic shortest path problems in networks with normally distributed link travel times with correlation. This algorithm scales efficiently in large networks and performs competitively with state-of-the-art methods. This package is designed for researchers and practitioners in transportation engineering, logistics, and telecommunications in which these types of networks may arise.

The package is available under MIT License.

PulseAlgorithm documentation is organized in two sections:

Graph Structure: documentation for the graph data structure used.

Data Loading: documentation for the data loader for real transportation networks found in Transportation Networks

Util: documentation for utility functions.

Pulse Algorithm for Reliable Shortest Path Problems (PA-RSPPs): documentation for the PA-RSPPs algorithm applied to different settings.

Installation

To install PulseAlgorithm.jl, you can use the Julia package manager. Open the Julia REPL and run the following command:

pkg> add https://github.com/EstebanLeiva/PulseAlgorithm.jl

Graph

PulseAlgorithm.Link — Type

Link(cost::Float64, mean::Float64, variance::Float64, value::Float64)

Simple link structure.

Parameters

cost::Float64: Link cost.

mean::Float64: Link travel time mean.

variance::Float64: Link travel time variance.

value::Float64: Link value.

PulseAlgorithm.Node — Type

Node(name::String, links::Dict{Int, Link}, incoming_links::Dict{Int, Link})

Simple node structure.

The links dictionary stores the outgoing links where the key is the destination node index and the value is the link itself. The incoming_links dictionary stores the incoming links where the key is the source node index and the value is the link itself.

PulseAlgorithm.Graph — Type

Graph(nodes::Dict{Int, Node}, name_to_index::Dict{String, Int})

Simple graph structure.

The nodes are stored in a dictionary where the key is the node index and the value is the node itself. The nametoindex dictionary is used to map the node name to the node index.

PulseAlgorithm.create_node! — Function

create_node!(graph::Graph, name::String)

Create a new node in the graph.

PulseAlgorithm.find — Function

find(graph::Graph, name::String)

Find the node index given the node name.

PulseAlgorithm.find_or_add! — Function

find_or_add!(graph::Graph, name::String)

Find the node index given the node name. If the node does not exist, create a new node.

PulseAlgorithm.add_link! — Function

add_link!(graph::Graph, src_name::String, dst_name::String, cost::Float64, mean::Float64, variance::Float64)

Add a link between two nodes.

PulseAlgorithm.get_links_info — Function

get_links_info(graph::Graph)

Return a dictionary with the link information.

Dijkstra

PulseAlgorithm.dijkstra — Function

dijkstra(graph::Graph, target_node::Int, path::String)

Return a vector with the cost to reach the target node from all other nodes in the graph.

Priority Queue implementation of Dijkstra's algorithm.

PulseAlgorithm.dijkstra_between_nodes — Function

dijkstra_between_nodes(graph::Graph, start_node::Int, target_node::Int, type::String)

Return the shortest path between two nodes in the graph.

Priority Queue implementation of Dijkstra's algorithm between two nodes.

Data Loader

PulseAlgorithm.TAData — Type

TAData(network_name::String, number_of_zones::Int, number_of_nodes::Int, first_thru_node::Int, number_of_links::Int, start_node::Array{Int,1}, end_node::Array{Int,1}, capacity::Array{Float64,1}, link_length::Array{Float64,1}, free_flow_time::Array{Float64,1}, B::Array{Float64,1}, power::Array{Float64,1}, speed_limit::Array{Float64,1}, toll::Array{Float64,1}, link_type::Array{Int64,1})

Simple structure to store the traffic assignment data.

PulseAlgorithm.load_ta — Function

load_ta(network_data_file::String, network_name::String)

Load the traffic assignment network data.

See also load_graph_from_ta.

PulseAlgorithm.load_flowCost — Function

load_flowCost(flow_file_dir:: String)

Load the flow cost data.

See also load_graph_from_ta.

PulseAlgorithm.calculate_fft_coefficient — Function

calculate_fft_coefficient(ta_data::TAData)

Calculate the average freeflowtime/link_length coefficient.

This is used to replace zero entries in the freeflowtime field.

See also load_graph_from_ta.

PulseAlgorithm.load_graph_from_ta — Function

load_graph_from_ta(tntp_file_dir::String, flow_file_dir:: String, network_name::String, CV::Float64, toll_factor::Float64, length_factor::Float64)

Load a graph with the traffic assignment data.

The cost, mean, and variance are calculated as explained in the Transportation Networks repository.

See also load_ta, load_flowCost, calculate_fft_coefficient.

Util

PulseAlgorithm.get_path_distribution — Function

get_path_distribution(graph::Graph, path::Vector{Int}, cov_dict::DefaultDict{Tuple{Int, Int, Int, Int}, Float64})

Return the mean, variance and covariance term of a given path's travel time distribution.

See also get_covariance_term.

PulseAlgorithm.modified_dfs — Function

modified_dfs(graph::Graph, start_link::Tuple{Int, Int}, max_depth::Int, depth::Int, visited_pairlinks::Dict{Tuple{Int, Int}, Int}, previous_node::Int)

Return the links at a distance less or equal to max_depth from the starting link.

See also get_covariance_dict.

PulseAlgorithm.get_covariance_dict — Function

get_covariance_dict(graph::Graph, ρ::Float64, max_depth::Int)

Return the covariance dictionary of a graph following a spatial correlation structure of radius=max_depth.

See also modified_dfs.

PulseAlgorithm.get_higher_priority_paths — Function

get_higher_priority_paths(pq::PriorityQueue, element::Vector{Int})

Return the elements with higher priority than the given element in a priority queue.

PulseAlgorithm.get_covariance_term — Function

get_covariance_term(covariance_dict::DefaultDict{Tuple{Int, Int, Int, Int}, Float64}, reachable_node::Int, path::Vector{Int})

Return the covariance term of the given path's travel time distribution.

See also get_path_distribution, pulse.

PA-S-$\alpha$RP

PulseAlgorithm.PaSarp — Type

PaSarp(G::Graph, α::Float64, covariance_dict::DefaultDict{Tuple{Int, Int, Int, Int}, Float64}, minimum_costs::Vector{Float64}, variance_costs::Vector{Float64}, mean_costs::Vector{Float64}, optimal_path::Vector{Int}, B::Float64, T_max::Float64, source_node::Int, target_node::Int, instance_info::Dict{String, Any})

Define the struct that will contain the information of the PA-S-αRP algorithm.

Parameters

G::Graph: Graph.

α::Float64: Reliability constraint.

covariance_dict::DefaultDict{Tuple{Int, Int, Int, Int}, Float64}: Covariance information.

minimum_costs::Vector{Float64}: Minimum cost to reach the target node.

variance_costs::Vector{Float64}: Minimum variance to reach the target node.

mean_costs::Vector{Float64}: Minimum mean to reach the target node.

optimal_path::Vector{Int}: Optimal path.

B::Float64: Best cost found so far (corresponds to the cost of the optimal path).

T_max::Float64: Travel time budget.

source_node::Int: Source node.

target_node::Int: Target node.

instance_info::Dict{String, Any}: Instance information.

PulseAlgorithm.initialize_PaSarp — Function

initialize_PaSarp(G::Graph, α::Float64, covariance_dict::DefaultDict{Tuple{Int, Int, Int, Int}, Float64}, source_node::String, target_node::String, T_max::Float64)

Initialize the PaSarp struct.

Arguments

G::Graph: Graph.

α::Float64: Reliability constraint.

covariance_dict::DefaultDict{Tuple{Int, Int, Int, Int}, Float64}: Covariance information.

source_node::String: Source node.

target_node::String: Target node.

T_max::Float64: Travel time budget.

PulseAlgorithm.preprocess! — Method

preprocess!(pa::PaSarp)

Calculate the minimum costs, variance costs, and mean costs to reach the target node; while ensuring that there is a path between the source and target nodes.

Note that this information could be saved into a file if the graph is static and could be reused in future runs.

PulseAlgorithm.check_feasibility — Method

check_feasibility(pa::PaSarp, current_node::Int, mean_path::Float64, variance_path::Float64, covariance_term_path::Float64, path::Vector{Int})

Check if the path is feasible with respect to the time budget and the reliability constraint.

Return true if the path is feasible; and false otherwise

PulseAlgorithm.check_bounds — Method

check_bounds(pa::PaSarp, current_node::Int, cost::Float64, path::Vector{Int})

Check if the partial path can improve the current solution.

Return true if it can improve the current solution; and false otherwise.

PulseAlgorithm.pulse — Method

pulse(pa::PaSarp, current_node::Int, cost::Float64, mean_path::Float64, variance_path::Float64, covariance_term_path::Float64, path::Vector{Int})

Propagate the pulses through the graph while checking the pruning criteria.

PulseAlgorithm.run_pulse — Method

run_pulse(pa::PaSarp, optimal_path::Vector{Int}, B::Float64)

Run the PA-S-αRP algorithm. Using an initial optimal path and cost is optional and it can be used to improve the performance of the algorithm.

Arguments

pa::PaSarp: PaSarp struct.

optimal_path::Vector{Int}=Vector{Int}(): Initial path.

B::Float64=Inf: Cost of the initial path.

PA-SD_RSPP

PulseAlgorithm.PaSdrspp — Type

PaSdrspp

A struct to store the information of the PaSdrspp algorithm.

Parameters

G::Graph: Graph.

α::Float64: Input reliability.

covariance_dict::DefaultDict{Tuple{Int, Int, Int, Int}, Float64}: Covariance information.

variance_costs::Vector{Float64}: Minimum variance to reach the target node.

mean_costs::Vector{Float64}: Minimum mean to reach the target node.

optimal_path::Vector{Int}: Optimal path.

B::Float64: Best quantile found so far (corresponds to the α-quantile of the optimal path).

source_node::Int: Source node.

target_node::Int: Target node.

PulseAlgorithm.initialize_PaSdrspp — Function

initialize_PaSdrspp(G::Graph, α::Float64, covariance_dict::DefaultDict{Tuple{Int, Int, Int, Int}, Float64}, source_node::String, target_node::String)

Initialize the PaSdrspp struct.

PulseAlgorithm.preprocess! — Method

preprocess!(sdp::PaSdrspp)

Preprocess the graph to calculate the minimum mean and variance to reach the target node.

Note that this information could be saved into a file if the graph is static and could be reused in future runs.

PulseAlgorithm.check_bounds — Method

check_bounds(sdp::PaSdrspp, current_node::Int, mean_path::Float64, variance_path::Float64, covariance_term_path::Float64, path::Vector{Int})

Check if the partial path can improve the current solution.

Return true if it can improve the current solution; and false otherwise.

PulseAlgorithm.pulse — Method

pulse(sdp::PaSdrspp, current_node::Int, cost::Float64, mean_path::Float64, variance_path::Float64, covariance_term_path::Float64, path::Vector{Int})

Propagate the pulses through the graph while checking the pruning criteria.

PulseAlgorithm.run_pulse — Method

run_pulse(sdp::PaSdrspp, optimal_path = Vector{Int}(), B = Inf)

Run the PA-SD-RSPP algorithm. Using an initial path and quantile is optional and it can be used to improve the performance of the algorithm.

Arguments

sdp::PaSdrspp: PaSdrspp struct.

optimal_path::Vector{Int}: Initial path.

B::Float64: Best quantile found so far (corresponds to the α-quantile of the initial path).